More refactorizations related to audio and course modules: - Use size_t instead of int - Removed several unused variables - Removed garbage whitespaces
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@310 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ game_ctrl.h
@@ -25,13 +25,13 @@ GNU General Public License for more details.
 
 struct TRace2 {
 	string race;
-	int course;
+	size_t course;
 	int light;
 	int snow;
 	int wind;
 	TIndex3 herrings;
 	TVector3 time;
-	int music_theme;
+	size_t music_theme;
 };
 
 struct TCup2 {
